+++ src/connected_socket.h
@@ -31,13 +31,13 @@ public:
 	 * Construct the connected socket object, given an open
 	 * socket.
 	 */
-	connected_socket_c( int socket, int port );
+	connected_socket_c( int socket, short port );
 	virtual ~connected_socket_c();
 
 	/**
 	 * Get the port that this socket was connected to.
 	 */
-	virtual int port() const { return m_port; }
+	virtual short port() const { return m_port; }
 
 	/**
 	 * Read a line of input from this connection.
@@ -57,7 +57,7 @@ public:
 
 private:
 	int m_socket;
-	int m_port;
+	short m_port;
 	line_parser_c m_line_parser;
 };
